examples/python/crewai-agent/requirements.txt
crewai>=0.80.0
python-dotenv>=1.0.0

traceroot==0.1.0
openinference-instrumentation-crewai>=0.1.0
